+++ fetch-rewards/app/authscreen.css
@@ -1,0 +1,113 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  align-items: center;
+  width: 100%;
+  height: 100vh;
+  background-color: cadetblue;
+
+  @media screen and (max-width: 1180px) {
+    text-align: center;
+    justify-content: inherit;
+  }
+}
+
+.moto {
+  position: fixed;
+  align-self: flex-start;
+  left: 20px;
+
+  @media screen and (max-width: 1180px) {
+    z-index: 1;
+    margin: 10% auto 0 auto;
+    position: relative;
+  }
+
+  @media screen and (max-width: 460px) {
+    left: auto;
+  }
+}
+
+.main {
+  z-index: 1;
+  width: 500px;
+  color: #4a5055;
+  
+  @media screen and (max-width: 1180px) {
+    margin-top: 30px;
+    width: 90%;
+  }
+
+  @media screen and (max-width: 460px) {
+    min-width: 372px;
+  }
+}
+
+.auth-container {
+  background-color: rgb(250, 235, 215);
+  max-width: 500px;
+  max-height: 500px;
+  width: 500px;
+  height: 500px;
+  margin: 0 auto;
+  padding: 40px;
+
+  @media screen and (max-width: 1180px) {
+    margin-top: 30px;
+    width: 90%;
+  }
+}
+
+.bgm {
+  position: fixed;
+  width: auto;
+  height: auto;
+  border-left: 2px solid white;
+  align-self: flex-end;
+}
+
+form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-evenly;
+  border: 1px solid white;
+  padding: 20px;
+  height: 100%;
+}
+
+h3 {
+  color: cadetblue;
+}
+
+label {
+  width: 80%;
+}
+
+.inputs {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+input {
+  margin: 6px 0 12px 0;
+  padding: 10px 12px;
+  width: 80%;
+  background-color: transparent;
+  color: #4a5055;
+  border: 1px solid #9f9f9f66;
+  box-shadow: 2px 2px 3px #98979791;
+}
+
+button {
+  padding: 10px 12px;
+  font-weight: 600;
+  background-color: #4a5055;
+  border-radius: .25rem;
+}
+
+.footer {
+}
