+++ fetch-rewards/app/page.js
@@ -1,95 +1,31 @@
 import Image from "next/image";
-import styles from "./page.module.css";
+import Form from 'next/form'
+import huskybgi from "../public/husky.png";
+import "./authscreen.css";
 
-export default function Home() {
+export default function AuthScreen() {
   return (
-    <div className={styles.page}>
-      <main className={styles.main}>
-        <Image
-          className={styles.logo}
-          src="/next.svg"
-          alt="Next.js logo"
-          width={180}
-          height={38}
-          priority
-        />
-        <ol>
-          <li>
-            Get started by editing <code>app/page.js</code>.
-          </li>
-          <li>Save and see your changes instantly.</li>
-        </ol>
-
-        <div className={styles.ctas}>
-          <a
-            className={styles.primary}
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              className={styles.logo}
-              src="/vercel.svg"
-              alt="Vercel logomark"
-              width={20}
-              height={20}
-            />
-            Deploy now
-          </a>
-          <a
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-            className={styles.secondary}
-          >
-            Read our docs
-          </a>
-        </div>
-      </main>
-      <footer className={styles.footer}>
-        <a
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/file.svg"
-            alt="File icon"
-            width={16}
-            height={16}
-          />
-          Learn
-        </a>
-        <a
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
-          Examples
-        </a>
-        <a
-          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
-          Go to nextjs.org →
-        </a>
-      </footer>
+    <div className="container">
+      <h2 className="moto">Bringing Families together</h2>
+      <div className="main">
+        <main className="auth-container">
+          <h2>Login / SignUp</h2>
+          <Form action="/search"> {/*//? This actions will need to be a function to auth and redirect on success //*/}
+            <h3>Please enter username and email for login or signup</h3>
+            <div className="inputs">
+              <label htmlFor="username">username</label>
+              <input id="username" name="username" />
+              <label htmlFor="email">email</label>
+              <input id="email" name="email" />
+            </div>
+            <button type="submit">Submit</button>
+          </Form>
+        </main>
+        <footer className={"footer"}>
+          © 2025 Vercel, Inc.
+        </footer> 
+      </div>
+      <Image className="bgm" src={huskybgi} alt="husky"/>
     </div>
   );
 }
